一次改来改去改错了后的提交，数据显示不出来

+++ app/src/main/java/com/example/abcdialogue/Weibo/Model/DTO/WBAllDTO.kt
@@ -10,7 +10,7 @@ data class WBAllDTO(
     @SerializedName("hasvisible")
     val hasvisible: Boolean?,
     @SerializedName("previous_cursor")
-    val previousCursor: Int?,
+    val previousCursor: Long?,
     @SerializedName("next_cursor")
     val nextCursor: Long?,
     @SerializedName("previous_cursor_str")
